Crate noir

TODO:

describe what noir is for

explain briefly the internals (job graph, execution graph, scheduler, blocks, operators, timestamps, watermarks, …)

simple examples

Re-exports

pub use config::EnvironmentConfig;

Modules

config

The types for constructing the configuration of the environment.

operator

Operators that can be applied to a stream.

structure

This mod contains the types and function that model the structure of the blocks and operators for debugging purposes only.

Structs

ExecutionMetadata

Metadata associated to a block in the execution graph.

IterationStateHandle

Handle to the state of the iteration.

KeyedStream

A KeyedStream is like a set of Streams, each of which partitioned by some Key. Internally it’s just a stream whose elements are KeyValue pairs and the operators behave following the KeyedStream semantics.

KeyedWindowedStream

A KeyedWindowedStream is a data stream partitioned by Key, where elements of each partition are divided in groups called windows. Each element can be assigned to one or multiple windows.

Stream

A Stream represents a chain of operators that work on a flow of data. The type of the elements that is leaving the stream is Out.

StreamEnvironment

Streaming environment from which it’s possible to register new streams and start the computation.

WindowedStream

A WindowedStream is a data stream where elements are divided in multiple groups called windows. Internally, a WindowedStream is just a KeyedWindowedStream where each element is assigned to the same key ().

Enums

BatchMode

Which policy to use for batching the messages before sending them.

Type Definitions

KeyValue

On keyed streams, this is the type of the items of the stream.