Struct noir::HttpResponse

pub struct HttpResponse<E: HttpEndpoint> {
    // some fields omitted
}

A mocked HTTP response that is being provided to a testable API.

Methods

impl<E: HttpEndpoint> HttpResponse<E>

fn with_status(self, status_code: StatusCode) -> Self

Sets the response status.

fn with_headers(self, headers: Vec<HttpHeader>) -> Self

Sets additional headers to be send with the response.

Use the headers![...] macro to easily create a vector containing concrete types of the hyper::Header trait for use with this method.

fn with_header<H: Header + HeaderFormat>(self, header: H) -> Self

Sets one additional header to be send with the response.

fn with_query(self, query: HttpQueryString) -> Self

Sets the query string of the response's path.

This will override any existing query string previously set or derived from the request's path.

fn with_body<S: Into<HttpBody>>(self, body: S) -> Self

Sets the response body.

fn with_error(self, error: Error) -> Self

Sets a low level io error to be returned once the response is read.

fn expected_header<H: Header + HeaderFormat>(self, header: H) -> Self

Sets one additional header that should be present on the request to the response.

Test Failure

If the header is either missing from the request or its value does not match the expected one.

fn unexpected_header<H: Header + HeaderFormat>(self) -> Self

Sets one additional header that should be absent from the request made to the response.

Test Failure

If the header is present on the request.

fn expected_headers(self, headers: Vec<HttpHeader>) -> Self

Sets additional headers that should be present on the request to the response.

Use the headers![...] macro to easily create a vector containing concrete types of the hyper::Header trait for use with this method.

Test Failure

If one or more of the headers are either missing from the request or their values differ from the expected ones.

fn expected_body<S: Into<HttpBody>>(self, body: S) -> Self

Sets the expected request body for the response.

The expected and the actual body are compared based on the MIME type of the reponse.

text/*

These Compared as strings, if no other charset is set in the response MIME type, UTF-8 will be used as the default encoding.

application/json

JSON objects are deep compared, but additional keys on response objects are ignored.

This allows for simpler and more fine grained assertions against JSON responses.

All other mime types

These are compared on a byte by byte basis.

Test Failure

If the actual response body does not match the expected one.

fn expected_exact_body<S: Into<HttpBody>>(self, body: S) -> Self

Sets the expected request body for the response, exact version.

This method is based on HttpRequest::expected_body() but performs additional comparison based on the mime type of the reponse:

application/json

In contrast to HttpResponse::expected_body() additional keys on response objects are compared and will fail the test.

All other mime types

See HttpResponse::expected_body().

Test Failure

If the actual response body does not match the expected one.

fn dump(self) -> Self

Dumps the response headers and body for this response.

Test Failure

Always.