Struct noir::MockResponseProvider

pub struct MockResponseProvider;

An interface for MockRequest trait implementations allowing them to consume matching MockResponse objects that were provided for the current test.

Methods

impl MockResponseProvider

fn response_from_request(request: Box<MockRequest>) -> Result<MockRequestResponse, Error> where Self: Sized

Returns the first, not yet consumed MockResponse that matches the provided request.

Matching is performed via MockResponse::matches_request().

Errors

When no matching response for the request exists.