Struct noir::Options

pub struct Options {
    pub json_compare_depth: usize,
}

Additional configuration options for api requests and responses.

Fields

json_compare_depth

Maximum depth for recursive json comparison. Defaults to 4096.

For top level json objects their own keys and the types of their values will be at depth 0, the actual values will then be at depth 1.

For top level json arrays, their length will be at depth 0 and their items and their types and values will be at level 1.