Struct noir::Request

pub struct Request<A: Api> {
    // some fields omitted
}

A HTTP used for testing APIs.

Methods

impl<A: Api> Request<A>

fn dump(self) -> Self

Dump the response header fields and body (if any) for debug purposes.

Note: Dumping will cause the test to panic.

fn with_headers<H: Header + HeaderFormat>(self, headers: Vec<H>) -> Self

Set additional request headers for this api request.

fn with_header<H: Header + HeaderFormat>(self, header: H) -> Self

Set one additional request header for this api request.

fn with_body<S: Into<RequestBody>>(self, body: S) -> Self

Set the body for this api request.

fn expect_status(self, status_code: StatusCode) -> Self

Set the expected response status for this api request.

fn expect_headers<H: Header + HeaderFormat>(self, headers: Vec<H>) -> Self

Set additional, expected response headers for this api request.

fn expect_header<H: Header + HeaderFormat>(self, header: H) -> Self

Set one additional, expected response header for this api request.

fn not_expect_headers<H: Header + HeaderFormat>(self, headers: Vec<H>) -> Self

Set additional, unexpected response headers for this api request.

fn not_expect_header<H: Header + HeaderFormat>(self, header: H) -> Self

Set one additional, unexpected response header for this api request.

fn expect_body<S: Into<RequestBody>>(self, body: S) -> Self

Set the expected response body for this request.

fn provide<T: MockResource + 'static>(self, resources: Vec<T>) -> Self

Provide mocked resource responses for this request.

fn mocks<T: MockProvider + 'static>(self, mocks: Vec<T>) -> Self

Provide custom mocks for this request.