Trait noir::Api

pub trait Api: Send + Copy + Default {
    fn hostname(&self) -> &'static str;
    fn port(&self) -> u16;
    fn start(&self);

    fn host(&self) -> String { ... }
    fn protocol(&self) -> &'static str { ... }
    fn url(&self) -> String { ... }
    fn url_with_path(&self, path: &str) -> String { ... }
    fn head(path: &'static str) -> HttpRequest<Self> where Self: 'static { ... }
    fn get(path: &'static str) -> HttpRequest<Self> where Self: 'static { ... }
    fn post(path: &'static str) -> HttpRequest<Self> where Self: 'static { ... }
    fn put(path: &'static str) -> HttpRequest<Self> where Self: 'static { ... }
    fn patch(path: &'static str) -> HttpRequest<Self> where Self: 'static { ... }
    fn delete(path: &'static str) -> HttpRequest<Self> where Self: 'static { ... }
}

A trait for the description of a testable API.

Required Methods

fn hostname(&self) -> &'static str

Returns the hostname of the API.

fn port(&self) -> u16

Returns the port of the API.

fn start(&self)

A blocking callback for starting the API.

The callback is executed in a background thread and should serve the API at the specified host.

Provided Methods

fn host(&self) -> String

Returns the hostname:port combination of the API used for making requests against the api.

fn protocol(&self) -> &'static str

Returns the protocol of the API.

By default this is based on the port, returning https for port 443.

fn url(&self) -> String

Returns the fully qualified base url of the API.

fn url_with_path(&self, path: &str) -> String

Returns the fully qualified url of the API with the specified path appended.

fn head(path: &'static str) -> HttpRequest<Self> where Self: 'static

Returns a HEAD request against the API.

fn get(path: &'static str) -> HttpRequest<Self> where Self: 'static

Returns a GET request against the API.

fn post(path: &'static str) -> HttpRequest<Self> where Self: 'static

Returns a POST request against the API.

fn put(path: &'static str) -> HttpRequest<Self> where Self: 'static

Returns a PUT request against the API.

fn patch(path: &'static str) -> HttpRequest<Self> where Self: 'static

Returns a PATCH request against the API.

fn delete(path: &'static str) -> HttpRequest<Self> where Self: 'static

Returns a DELETE request against the API.