Trait noir::HttpApi

pub trait HttpApi: Send + Copy + Default {
    fn start(&self);
    fn hostname(&self) -> &'static str;
    fn port(&self) -> u16;

    fn host(&self) -> String { ... }
    fn protocol(&self) -> &'static str { ... }
    fn url(&self) -> String { ... }
    fn url_with_path(&self, path: &str) -> String { ... }
    fn options(path: &'static str) -> HttpRequest<Self> where Self: 'static { ... }
    fn get(path: &'static str) -> HttpRequest<Self> where Self: 'static { ... }
    fn post(path: &'static str) -> HttpRequest<Self> where Self: 'static { ... }
    fn put(path: &'static str) -> HttpRequest<Self> where Self: 'static { ... }
    fn delete(path: &'static str) -> HttpRequest<Self> where Self: 'static { ... }
    fn head(path: &'static str) -> HttpRequest<Self> where Self: 'static { ... }
    fn trace(path: &'static str) -> HttpRequest<Self> where Self: 'static { ... }
    fn connect(path: &'static str) -> HttpRequest<Self> where Self: 'static { ... }
    fn patch(path: &'static str) -> HttpRequest<Self> where Self: 'static { ... }
    fn ext(http_verb: &'static str, path: &'static str) -> HttpRequest<Self> where Self: 'static { ... }
}

A trait for the description of a testable, HTTP based api.

Example Implementation

use noir::HttpApi;

#[derive(Copy, Clone, Default)]
struct Api;
impl HttpApi for Api {

    fn hostname(&self) -> &'static str {
        "localhost"
    }

    fn port(&self) -> u16 {
        8080
    }

    fn start(&self) {
        // Start the HTTP server...
    }

}

Required Methods

fn start(&self)

A blocking callback for provision of the api.

The callback is executed in a background thread and should serve the api at the specified host.

fn hostname(&self) -> &'static str

Returns the hostname of the api.

fn port(&self) -> u16

Returns the port of the api.

Provided Methods

fn host(&self) -> String

Returns the hostname:port combination of the api used for making requests against the api.

fn protocol(&self) -> &'static str

Returns the http protocol used by the api.

By default this is based on the port, returning https for port 443.

fn url(&self) -> String

Returns the fully qualified base url of the api.

fn url_with_path(&self, path: &str) -> String

Returns the fully qualified url of the api with the specified path appended.

fn options(path: &'static str) -> HttpRequest<Self> where Self: 'static

Returns a OPTIONS request that will be performed against the api.

fn get(path: &'static str) -> HttpRequest<Self> where Self: 'static

Returns a GET request that will be performed against the api.

fn post(path: &'static str) -> HttpRequest<Self> where Self: 'static

Returns a POST request that will be performed against the api.

fn put(path: &'static str) -> HttpRequest<Self> where Self: 'static

Returns a PUT request that will be performed against the api.

fn delete(path: &'static str) -> HttpRequest<Self> where Self: 'static

Returns a DELETE request that will be performed against the api.

fn head(path: &'static str) -> HttpRequest<Self> where Self: 'static

Returns a HEAD request that will be performed against the api.

fn trace(path: &'static str) -> HttpRequest<Self> where Self: 'static

Returns a TRACE request that will be performed against the api.

fn connect(path: &'static str) -> HttpRequest<Self> where Self: 'static

Returns a CONNECT request that will be performed against the api.

fn patch(path: &'static str) -> HttpRequest<Self> where Self: 'static

Returns a PATCH request that will be performed against the api.

fn ext(http_verb: &'static str, path: &'static str) -> HttpRequest<Self> where Self: 'static

Returns a request for the specified http verb extension that will be performed against the api.